+++ packages/apollo-datasource-rest/src/__tests__/RESTDataSource.test.ts
@@ -639,7 +639,7 @@ describe('RESTDataSource', () => {
     it('allows disabling the GET cache', async () => {
       const dataSource = new (class extends RESTDataSource {
         override baseURL = 'https://api.example.com';
-        override requestCacheEnabled = false;
+        override memoizeGetRequests = false;
 
         getFoo(id: number) {
           return this.get(`foo/${id}`);
@@ -808,7 +808,7 @@ describe('RESTDataSource', () => {
     it('allows setting cache options for each request', async () => {
       const dataSource = new (class extends RESTDataSource {
         override baseURL = 'https://api.example.com';
-        override requestCacheEnabled = false;
+        override memoizeGetRequests = false;
 
         getFoo(id: number) {
           return this.get(`foo/${id}`);
@@ -837,7 +837,7 @@ describe('RESTDataSource', () => {
     it('allows setting a short TTL for the cache', async () => {
       const dataSource = new (class extends RESTDataSource {
         override baseURL = 'https://api.example.com';
-        override requestCacheEnabled = false;
+        override memoizeGetRequests = false;
 
         getFoo(id: number) {
           return this.get(`foo/${id}`);